git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Incubation/SimuLizar/branches@31626 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ Elasticity and RR model merge/org.palladiosimulator.simulizar/src/org/palladiosimulator/simulizar/runtimestate/SimulatedBasicComponentInstance.java
@@ -28,7 +28,7 @@ public class SimulatedBasicComponentInstance extends SimulatedComponentInstance 
         final AssemblyContext myAssCtx = fqID.getAssembyContextPath().get(fqID.getAssembyContextPath().size() - 1);
         for (final PassiveResource passiveResource : passiveResources) {
             final long initialCount = (long) StackContext.evaluateStatic(
-                    passiveResource.getCapacity_PassiveResource().getSpecification(), Integer.class,
+                    passiveResource.getCapacity_PassiveResource().getSpecification(), Long.class,
                     context.getStack().currentStackFrame());
             final IPassiveResource simulatedResource = new SimSimpleFairPassiveResource(passiveResource, myAssCtx,
                     this.getRuntimeState().getModel(), initialCount);
